feat(web): move Appearance + Demo mode into a dedicated Settings › Appearance tab
Split the Appearance card and the per-device Demo mode card out of the
Account & tokens page into a new /settings/appearance page (mirrors the
Run defaults split), inserted as the second Settings tab. Behavior of the
theme/mode/typeface controls and the demo toggle is unchanged; only their
home moved. Drops the now-unused helpers/imports from Settings.tsx.
