+++ web/src/components/SettingsShell.tsx
@@ -15,6 +15,7 @@ import { cx, PageHeader } from "./ui";
 
 const TABS = [
   { to: "/settings", label: "Account & tokens", end: true },
+  { to: "/settings/appearance", label: "Appearance", end: false },
   { to: "/settings/run-defaults", label: "Run defaults", end: false },
   { to: "/settings/forge", label: "Forge", end: false },
   { to: "/settings/access", label: "Access", end: false },
@@ -30,10 +31,11 @@ export function SettingsShell({ description, children }: { description: string; 
       <PageHeader title="Settings" description="Your personal uzi configuration." />
       {/* Issue #204: the tab row scrolls WITHIN its own container so the page body never
           scrolls horizontally. Measured when this row held five tabs: at 390px the strip
-          overflowed (scrollWidth 401 vs clientWidth 390). Four tabs likely fit today, but
-          `overflow-x-auto` stays so the next added tab cannot regress the page body.
-          The tabs are `shrink-0`/`whitespace-nowrap` so they overflow-and-scroll rather
-          than compress to fit (which would defeat overflow-x). */}
+          overflowed (scrollWidth 401 vs clientWidth 390). The row now holds six tabs
+          (Appearance was added), so `overflow-x-auto` earns its keep — it lets the strip
+          scroll rather than forcing the page body sideways. The tabs are
+          `shrink-0`/`whitespace-nowrap` so they overflow-and-scroll rather than compress
+          to fit (which would defeat overflow-x). */}
       <div className="flex gap-1 overflow-x-auto border-b border-edge">
         {TABS.map((t) => (
           <NavLink
